Made tasks to render as table rows instead of separate items. Renamed component TaskItem to TaskRow to better represent internal content.
